Make captured content available
One may now use regex/4 to access the content captured by parentheses in
a pattern.  Because regex/4 is a strict superset of =~/2 and \~/2, those
two operators are now implemented in terms of the former.  This
simplifies the code quite a bit.  It should also make it behave more
consistently in the long run since the operators are just sugar for the
underlying, verbose predicate.

Closes #1