TODO test for excess capture names
regex/4 named capture support is designed to facilitate
prolog_load_context(variable_names,_)  Eventually =~ will be desugared
into a call to regex/4 but there will almost always be more variables in
scope than the ones mentioned in named captures in a regex.

This test describes the behavior we want in that situation.  Namely, the
superfluous variables are left untouched.