Moved ontrace and static_strip_module to xlibrary, to avoid dependencies of rtchecks and assertions.
Now the libraries are collected and loaded on the fly.

Added option checkload, to check that the dependencies between packages are
correct.

moved predicate_from to its own module, to avoid dependencies to xtools from
packages of lower hierarchy

moved clause of extra_clauses to ai_extra_clauses.pl to avoid dependencies of
xlibrary with xtools: that will not be necessary in the future.